moved "check repository" to "deploy"
